# shareholdergame-facade/src/main/java/com/shareholdergame/engine/facade/dto/SignUp.java
package com.shareholdergame.engine.facade.dto;

public class SignUp {

    private String userName;

    private String email;

    private String password;

    public String getUserName() {
        return userName;
    }

    public void setUserName(String userName) {
        this.userName = userName;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}