# shareholdergame-facade/src/main/java/com/shareholdergame/engine/facade/dto/player/Player.java
package com.shareholdergame.engine.facade.dto.player;

import java.io.Serializable;

/**
 * Date: 10/08/2018
 *
 * @author Aliaksandr Savin
 */
public class Player implements Serializable {

    private String id;

    private String name;

    private String avatarUrl;

    private boolean online;

    private boolean removed;

    private boolean friend;

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getAvatarUrl() {
        return avatarUrl;
    }

    public void setAvatarUrl(String avatarUrl) {
        this.avatarUrl = avatarUrl;
    }

    public boolean isOnline() {
        return online;
    }

    public void setOnline(boolean online) {
        this.online = online;
    }

    public boolean isRemoved() {
        return removed;
    }

    public void setRemoved(boolean removed) {
        this.removed = removed;
    }

    public boolean isFriend() {
        return friend;
    }

    public void setFriend(boolean friend) {
        this.friend = friend;
    }
}